Oppg2/requirements.txt
mysql-connector==2.2.9
pyFirmata==1.1.0
pyserial==3.5
